requirements.txt
python-dotenv>=0.21.0
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.4
langchain-community>=0.0.5
duckduckgo-search>=2.6.1
requests>=2.31.0
